requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
passlib==1.7.4
python-multipart==0.0.5
PyJWT==2.1.0
bcrypt==3.2.0
pymongo==4.6.0